requirements.txt
# ApexAI-Core Dependencies
requests>=2.31.0
uvicorn>=0.23.0
fastapi>=0.103.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

